pyproject.toml
[project]
name = "doc-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "camelot-py[cv]>=1.0.9",
    "googlenews>=1.6.15",
    "newsapi-python>=0.2.7",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "yfinance>=0.2.66",
]
